Structure: Update and move Common components to own directory. NavBar updated to use constants from lang/config to create menu items and links. Fast hack to resolve language change functionality.
