+++ src/components/Common/Footer.jsx
@@ -1,5 +1,5 @@
-import { h } from 'preact';
-import Styles from './styles.module.scss';
+import { h } from "preact";
+import Styles from "./Footer.module.scss";
 
 function Footer() {
   return (
