+++ src/components/Common/Header.jsx
@@ -1,0 +1,6 @@
+import { h } from "preact";
+function Header() {
+  return <header></header>;
+}
+
+export default Header;
