+++ src/components/Common/Navbar.module.scss
@@ -2,12 +2,12 @@
 .nav {
   display: flex;
   align-items: center;
-  padding-top: 2.75rem;
-  padding-right: 1rem;
+  padding-top: 2.5rem;
+  padding-right: 1.5rem;
   padding-bottom: 1.5rem;
-  padding-left: 1rem;
+  padding-left: 2rem;
   @media (min-width: $min-mobile) {
-    padding-left: 2rem;
+    padding: 2.5rem 3rem 2.5rem 4rem;
   }
 }
 
@@ -37,6 +37,18 @@
   &:focus,
   &:hover {
     color: $t-active;
+  }
+}
+
+.langSwitch {
+  display: flex;
+  flex-direction: column;
+  .lang {
+    padding: 0.35em;
+    border: 1px rgba(0, 0, 0, 35);
+    + .lang {
+      margin-left: -1px;
+    }
   }
 }
