+++ src/components/Common/index.js
@@ -1,0 +1,5 @@
+import Navbar from "./Navbar.jsx";
+import Header from "./Header.jsx";
+import Footer from "./Footer.jsx";
+
+export { Navbar, Header, Footer };
