+++ src/components/MainHead.astro
@@ -1,13 +1,17 @@
 ---
-const { title } = Astro.props;
+import {SITE} from '$/config.js';
+const { content = {}, canonicalURL} = Astro.props;
 ---
+ 
 
 <meta charset="UTF-8">
 <meta name="viewport" content="width=device-width, initial-scale=1" />
-<meta name="title" property="og:type" content="Pablo Anttila — Frontend Developer specialized in JavaScript for Web and Mobile." />
+<meta name="title" property="og:title" content={SITE.title} />
 <meta name="type" property="og:type" content="website" />
-<meta name="url" property="og:url" content="https://papplo.github.io"/>
-<meta name="description" property="og:description" 
-content="Connect with Pablo Anttila, discover recent projects and blog posts. This portfolio site was build using Astro SSG and React deployed with CI tools to Github Pages."/>
-<title>{title}</title>
+<meta name="url" property="og:url" content={canonicalURL}/>
+<meta property="og.locale" content={SITE.defaultLanguage}/>
+<meta property="og:description" content={content.description ? content.description : SITE.description}/>
+<meta property="og:site_name" content={SITE.title}/>
+
+<title>{SITE.title}</title>
 <link rel="stylesheet" type="text/css" href="/styles/global.css">
